requirements.txt
# langchain
# langgraph
# tqdm
# openai
# streamlit
# google-search-results==2.4.2


langchain==0.1.14
langchain-community==0.0.30
openai==1.14.2
langgraph==0.0.19
streamlit==1.33.0
google-search-results==2.4.2
tqdm==4.66.2
python-dotenv==1.0.1